Dtos for most of entities, User CRUD

+++ Migrations/BookHubDbContextModelSnapshot.cs
@@ -240,8 +240,8 @@ namespace BookHub.Migrations
                     b.Property<bool>("IsDeleted")
                         .HasColumnType("INTEGER");
 
-                    b.Property<int>("Price")
-                        .HasColumnType("INTEGER");
+                    b.Property<double>("Price")
+                        .HasColumnType("REAL");
 
                     b.Property<string>("Publisher")
                         .IsRequired()
@@ -275,7 +275,7 @@ namespace BookHub.Migrations
                             Description = "Ve světě, který leží na krunýři obrovské želvy, se vydává na cestu rozverná, temperamentní a neuvěřitelně výstřední výprava. Setkáte se s lakomým a naprosto neschopným čarodějem Mrakoplašem, naivním turistou Dvoukvítkem, jehož Zavazadlo za ním běhá jako pes na stovce malých nožiček, s draky, kteří existují, pokud na ně opravdu věříte, a samozřejmě dojdete až na okraj této podivné planety.",
                             ISBN = "978-80-7197-614-1",
                             IsDeleted = false,
-                            Price = 399,
+                            Price = 399.0,
                             Publisher = "Talpress",
                             Quantity = 7,
                             ReleaseYear = 1993,
@@ -289,7 +289,7 @@ namespace BookHub.Migrations
                             Description = "Harry Potter je sirotek, který žije u svých příbuzných Dursleyových. Jeho rodiče byli mocní čarodějové, kteří zahynuli při souboji s nejtemnějším čarodějem všech dob, Lordem Voldemortem. Harryho rodiče zanechali svého syna v péči svého přítele, kouzelníka Albusa Brumbála, ředitele Školy čar a kouzel v Bradavicích. Harryho příbuzní o jeho magických schopnostech nevědí, protože se bojí, že by ho mohli zavděčit. Harryho život je plný ponižování a šikany, ale v den jeho jedenáctých narozenin se vše změní. Harry dostane dopis od Brumbála, který ho pozve na Školu čar a kouzel v Bradavicích. Harry se dozví, že je čaroděj a že jeho rodiče zemřeli při souboji s nejtemnějším čarodějem všech dob, Lordem Voldemortem. Harry se vydává do Bradavic, kde se seznámí s Ronem Weasleym a Hermionou Grangerovou, kteří se stanou jeho nejlepšími přáteli. Harry se také dozví, že Voldemort přežil a že se chystá získat Kámen mudrců, který mu pomůže znovu získat svou moc.",
                             ISBN = "978-80-7197-614-1",
                             IsDeleted = false,
-                            Price = 399,
+                            Price = 399.0,
                             Publisher = "Albatros",
                             Quantity = 17,
                             ReleaseYear = 1997,
@@ -303,7 +303,7 @@ namespace BookHub.Migrations
                             Description = "Když král Robert rozhodne, že jeho nejstarší přítel Eddard Stark bude jeho pravou rukou, nevědomky odstartuje události, které otřesou celým kontinentem. Eddard se totiž snaží vyšetřit tajemnou smrt předchozího krájů a nastoupit na své nové místo, ale brzy zjistí, že je vše mnohem složitější, než se zdálo. Na jihu se totiž připravuje vzpoura a v záloze číhá starodávné zlo, které se probouzí.",
                             ISBN = "978-80-257-2891-5",
                             IsDeleted = false,
-                            Price = 699,
+                            Price = 699.0,
                             Publisher = "Argo",
                             Quantity = 15,
                             ReleaseYear = 1996,
@@ -317,7 +317,7 @@ namespace BookHub.Migrations
                             Description = "V roce 1958 se sedmička přátel z Derry, malého městečka v americkém státě Maine, vydává do kanalizace, aby zničila zlého klauna Pennywaise, který se zde ukrývá. Sedmička však zjistí, že Pennywise je jen jedním z mnoha podob zla, které se ukrývá v Derry. Ze slibu se však stane kletba a sedmička se musí v roce 1985 vrátit do Derry, aby zlo zničila jednou provždy.",
                             ISBN = "978-80-7197-614-1",
                             IsDeleted = false,
-                            Price = 799,
+                            Price = 799.0,
                             Publisher = "Talpress",
                             Quantity = 2,
                             ReleaseYear = 1986,
@@ -331,7 +331,7 @@ namespace BookHub.Migrations
                             Description = "Harry Potter, Ron Weasley a Hermiona Grangerová se vrací do Bradavic, kde se dozvědí o Tajemné komnatě, která je úzce spojena s Harrym. Harry se rozhodne najít Tajemnou komnatu a zjistit, co se tam skrývá. Harryho plán je však zmařen, když se objeví záhadný písař. Harryho přítel Ron je napaden a jeho sestra Ginny zmizí. Harry, Ron a Hermiona se tak vydávají do Tajemné komnaty, aby zjistili, co se tam skrývá a zachránili Ginny.",
                             ISBN = "978-80-7197-614-1",
                             IsDeleted = false,
-                            Price = 399,
+                            Price = 399.0,
                             Publisher = "Albatros",
                             Quantity = 9,
                             ReleaseYear = 1998,
@@ -345,7 +345,7 @@ namespace BookHub.Migrations
                             Description = "Harry Potter se vrací do Bradavic, ale tentokrát se musí vyhýbat nebezpečnému vězni jménem Sirius Black, který utekl z Azkabanu. Harry se dozví, že Sirius byl vězněn kvůli tomu, že zradil jeho rodiče a že se chystá Harryho zabít. Harry se vydává na cestu, která ho zavede do minulosti, kde se dozví, že Sirius není tím, za koho se vydává.",
                             ISBN = "978-80-7197-614-1",
                             IsDeleted = false,
-                            Price = 399,
+                            Price = 399.0,
                             Publisher = "Albatros",
                             Quantity = 11,
                             ReleaseYear = 1999,
@@ -359,7 +359,7 @@ namespace BookHub.Migrations
                             Description = "Harry Potter se vrací do Bradavic, kde se má konat Turnaj tří kouzelníků. Harry se však dozví, že se do turnaje dostal podvodem a že se musí zúčastnit tří nebezpečných úkolů. Harry se vydává na cestu, která ho zavede do minulosti, kde se dozví, že se do turnaje dostal podvodem a že se musí zúčastnit tří nebezpečných úkolů.",
                             ISBN = "978-80-7197-614-1",
                             IsDeleted = false,
-                            Price = 399,
+                            Price = 399.0,
                             Publisher = "Albatros",
                             Quantity = 13,
                             ReleaseYear = 2000,
@@ -373,7 +373,7 @@ namespace BookHub.Migrations
                             Description = "Do Bradavic přišly temné časy. Po útoku mozkomorů na bratrance Dudleyho Harry ví, že Voldemort udělá cokoli, jen aby ho našel. Mnozí jeho návrat popírají, ale Harry přesto není sám: na Grimmauldově náměstí se schází tajný řád, který chce bojovat proti temným silám. Harry se musí od profesora Snapea naučit, jak se chránit před Voldemortovými útoky na jeho duši. Jenže Pán zla je den ode dne silnější a Harrymu dochází čas…",
                             ISBN = "978-80-7197-614-1",
                             IsDeleted = false,
-                            Price = 399,
+                            Price = 399.0,
                             Publisher = "Albatros",
                             Quantity = 13,
                             ReleaseYear = 2000,
@@ -387,7 +387,7 @@ namespace BookHub.Migrations
                             Description = "Moc Lorda Voldemorta stále roste a smrtijedi působí spoušť ve světě mudlů i kouzelníků. Když Harry Potter objeví starou učebnici lektvarů patřící tajemnému princi dvojí krve, spoléhá na její kouzla i přes varování svých kamarádů. Profesor Brumbál poodhaluje Voldemortovu minulost a s Harryho pomocí se snaží odkrýt tajemství jeho nesmrtelnosti. Jenže zlo se dere k moci stále silněji, neštěstí se blíží a Bradavice už nikdy nebudou jako dřív.",
                             ISBN = "978-80-7197-614-1",
                             IsDeleted = false,
-                            Price = 399,
+                            Price = 399.0,
                             Publisher = "Albatros",
                             Quantity = 13,
                             ReleaseYear = 2005,
@@ -401,7 +401,7 @@ namespace BookHub.Migrations
                             Description = "Harry Potter se vydává na nebezpečnou cestu, aby zničil poslední Voldemortovy viteály. Společně s Ronem a Hermionou hledá zbytek Voldemortovy duše, který se ukrývá v tělech jeho nejmocnějších stoupenců. Harry se musí vydat na nebezpečnou cestu, aby zničil poslední Voldemortovy viteály. Společně s Ronem a Hermionou hledá zbytek Voldemortovy duše, který se ukrývá v tělech jeho nejmocnějších stoupenců.",
                             ISBN = "978-80-7197-614-1",
                             IsDeleted = false,
-                            Price = 399,
+                            Price = 399.0,
                             Publisher = "Albatros",
                             Quantity = 13,
                             ReleaseYear = 2007,
@@ -415,7 +415,7 @@ namespace BookHub.Migrations
                             Description = "Harry Potter je zaměstnán v Ministerstvu kouzel a má tři školáky. Jeho minulost ho však neustále pronásleduje. Harry Potter je zaměstnán v Ministerstvu kouzel a má tři školáky. Jeho minulost ho však neustále pronásleduje.",
                             ISBN = "978-80-7197-614-1",
                             IsDeleted = false,
-                            Price = 399,
+                            Price = 399.0,
                             Publisher = "Albatros",
                             Quantity = 13,
                             ReleaseYear = 2016,
